Jenkinsfile refactored

+++ Jenkinsfile
@@ -6,14 +6,45 @@ pipeline {
     dockerHubImagetag = ''
     artifactoryImageLatest = ''
     artifactoryImageTag = ''
-    imageLine = 'datoma/teslamate:latest'
+
     GIT_URL = 'https://github.com/datoma/teslamate.git'
     GIT_BRANCH = 'master'
     GIT_CREDENTIALS = 'Github'
+    DOCKERHUB_URL = 'https://registry.hub.docker.com'
+    DOCKERHUB_CREDENTIALS = 'dockerhub'
+    ARTIFACTORY_URL = 'https://datoma.jfrog.io/artifactory'
+    ARTIFACTORY_CREDENTIALS = 'ArtifactoryDockerhub'
+
+    TRIVY_VERSION = 'datoma/trivy-server:0.15.0'
+    DOCKLE_VERSION = 'datoma/dockle:0.3.1'
+    HADOLINT_VERSION = 'hadolint/hadolint'
+  }
+
+  parameters{
+    string(defaultValue: "latest", description: 'tag to build/push', name: 'DOCKER_IMAGE_TAG')
+    booleanParam(defaultValue: false, description: 'create release branch', name: 'CREATE_RELEASE_BRANCH')
+    booleanParam(defaultValue: false, description: 'deploy to dockerhub', name: 'PUSH_DOCKER')
+    booleanParam(defaultValue: false, description: 'deploy to artifactory', name: 'PUSH_ARTIFACTORY')
+  }
+
+  options {
+    buildDiscarder(logRotator(numToKeepStr: '', artifactNumToKeepStr: '20'))
+    disableConcurrentBuilds()
   }
 
   agent any
   stages {
+    stage("check params") {
+      steps {
+        script {
+          params.each {
+            if (DOCKER_IMAGE_TAG == null || DOCKER_IMAGE_TAG == "" || DOCKER_IMAGE_TAG == "latest")
+              error "This pipeline stops here because no tag was set (var is ${DOCKER_IMAGE_TAG})"
+            }
+        }
+      }
+    }
+
     stage ("prepare") {
       steps {
         script {
@@ -37,47 +68,48 @@ pipeline {
       }
     }
 
+    stage('create branch') {
+      when {
+        expression { 
+          params.CREATE_RELEASE_BRANCH == true
+        }
+      }
+      steps {
+        withCredentials([sshUserPrivateKey(credentialsId: 'Github_ssh', keyFileVariable: '', passphraseVariable: '', usernameVariable: '')]) {
+          //sh("git push origin ${GIT_BRANCH}:release-${DOCKER_IMAGE_TAG}")
+          sh("git checkout -b releases/${DOCKER_IMAGE_TAG}")
+          sh("git push --set-upstream origin releases/${DOCKER_IMAGE_TAG}")
+        }
+       }
+    }
+
     stage('Building the image') {
       steps {
         script {
-          dockerHubImageLatest = docker.build("${DOCKERHUB_IMAGE_NAME}:latest")
+          docker.withRegistry("${DOCKERHUB_URL}", "${DOCKERHUB_CREDENTIALS}") {
+            dockerHubImageLatest = docker.build("${DOCKERHUB_IMAGE_NAME}:latest")
+          }
         }
       }
     }
     
     stage('Tagging the image') {
-      parallel {
-        stage('push Dockerhub Tag') {
-          steps {
-            script {
-              dockerHubImagetag = docker.build("${DOCKERHUB_IMAGE_NAME}:${DOCKER_IMAGE_TAG}")
-            }
-          }
-        }
-        stage('push Artifactory latest') {
-          steps {
-            script {
-              artifactoryImageLatest = docker.build("${ARTIFACTORY_IMAGE_NAME}:latest")
-            }
-          }
-        }
-        stage('push Artifactory Tag') {
-          steps {
-            script {
-              artifactoryImageTag = docker.build("${ARTIFACTORY_IMAGE_NAME}:${DOCKER_IMAGE_TAG}")
-            }
+      steps {
+        script {
+          docker.withRegistry("${DOCKERHUB_URL}", "${DOCKERHUB_CREDENTIALS}") {
+            dockerHubImagetag = docker.build("${DOCKERHUB_IMAGE_NAME}:${DOCKER_IMAGE_TAG}")
           }
         }
       }
     }
 
-    stage('Trivy, dockle and hadolint tests') {
+    stage('Test stages') {
       parallel {
         stage('Trivy Tag and latest') {
           steps {
               script {
-                trivy_latest = sh(returnStdout: true, script: 'docker run --name trivy-client --rm -i -v /var/run/docker.sock:/var/run/docker.sock:ro datoma/trivy-server:latest trivy client --remote https://trivy.blackboards.de ${DOCKERHUB_IMAGE_NAME}:latest')
-                trivy_tag = sh(returnStdout: true, script: 'docker run --name trivy-client --rm -i -v /var/run/docker.sock:/var/run/docker.sock:ro datoma/trivy-server:latest trivy client --remote https://trivy.blackboards.de --ignore-unfixed --exit-code 1 --severity CRITICAL,HIGH,MEDIUM ${DOCKERHUB_IMAGE_NAME}:${DOCKER_IMAGE_TAG}')
+                trivy_latest = sh(returnStdout: true, script: 'docker run --name trivy-client --rm -i -v /var/run/docker.sock:/var/run/docker.sock:ro ${TRIVY_VERSION} trivy client --remote https://trivy.blackboards.de ${DOCKERHUB_IMAGE_NAME}:latest')
+                trivy_tag = sh(returnStdout: true, script: 'docker run --name trivy-client --rm -i -v /var/run/docker.sock:/var/run/docker.sock:ro ${TRIVY_VERSION} trivy client --remote https://trivy.blackboards.de --ignore-unfixed --severity CRITICAL,HIGH,MEDIUM ${DOCKERHUB_IMAGE_NAME}:${DOCKER_IMAGE_TAG}')
               }
               echo "TRIVY latest: ${trivy_latest}"
               writeFile(file: 'trivy-latest.txt', text: "${trivy_latest}")
@@ -88,7 +120,7 @@ pipeline {
         stage('dockle Tag') {
           steps {
             script {
-                dockle_tag = sh(returnStdout: true, script: 'docker run --rm -v /var/run/docker.sock:/var/run/docker.sock datoma/dockle:latest ${DOCKERHUB_IMAGE_NAME}:${DOCKER_IMAGE_TAG}')
+                dockle_tag = sh(returnStdout: true, script: 'docker run --rm -v /var/run/docker.sock:/var/run/docker.sock ${DOCKLE_VERSION} ${DOCKERHUB_IMAGE_NAME}:${DOCKER_IMAGE_TAG}')
               }
               echo "Dockle tag: ${dockle_tag}"
               writeFile(file: 'dockle_tag.txt', text: "${dockle_tag}")
@@ -98,7 +130,7 @@ pipeline {
           steps {
             script {
               try {
-                sh 'docker run --rm -i hadolint/hadolint < Dockerfile | tee hadolint_tag.txt'
+                sh 'docker run --rm -i ${HADOLINT_VERSION} < Dockerfile | tee hadolint_tag.txt'
               } catch (err) {
                 echo err.getMessage()
               }
@@ -109,22 +141,27 @@ pipeline {
       }
     }
 
-    stage('Deploy the latest image') {
+    stage('Deploy Images to Dockerhub') {
+      when {
+        expression { 
+          params.PUSH_DOCKER == true
+        }
+      }
       parallel {
         stage('Deploy image with latest to Dockerhub') {
           steps {
             script {
-              docker.withRegistry('https://registry.hub.docker.com', 'dockerhub') {
+              docker.withRegistry("${DOCKERHUB_URL}", "${DOCKERHUB_CREDENTIALS}") {
                 dockerHubImageLatest.push()
               }
             }
           }
         }
-        stage('Deploy image with latest to Artifactory') {
+        stage('Deploy image with tag to Dockerhub') {
           steps {
             script {
-                docker.withRegistry('https://datoma.jfrog.io/artifactory', 'ArtifactoryDockerhub') {
-                artifactoryImageLatest.push()
+              docker.withRegistry("${DOCKERHUB_URL}", "${DOCKERHUB_CREDENTIALS}") {
+                dockerHubImagetag.push()
               }
             }
           }
@@ -132,21 +169,40 @@ pipeline {
       }
     }
 
-    stage('Deploy the tagged image') {
+    stage('Tagging Artifactory images') {
+      when {
+        expression { 
+          params.PUSH_ARTIFACTORY == true
+        }
+      }
+      steps {
+        script {
+          artifactoryImageLatest = docker.build("${ARTIFACTORY_IMAGE_NAME}:latest")
+          artifactoryImageTag = docker.build("${ARTIFACTORY_IMAGE_NAME}:${DOCKER_IMAGE_TAG}")
+        }
+      }
+    }
+
+    stage('Push images to Artifactory') {
+      when {
+        expression { 
+          params.PUSH_ARTIFACTORY == true
+        }
+      }
       parallel {
-        stage('Deploy image with tag to Dockerhub') {
+        stage('Push image with latest to Artifactory') {
           steps {
             script {
-              docker.withRegistry('https://registry.hub.docker.com', 'dockerhub') {
-                dockerHubImagetag.push()
+                docker.withRegistry("${ARTIFACTORY_URL}", "${ARTIFACTORY_CREDENTIALS}") {
+                artifactoryImageLatest.push()
               }
             }
           }
         }
-        stage('Deploy image with tag to Artifactory') {
+        stage('Push image with tag to Artifactory') {
           steps {
             script {
-                docker.withRegistry('https://datoma.jfrog.io/artifactory', 'ArtifactoryDockerhub') {
+              docker.withRegistry("${ARTIFACTORY_URL}", "${ARTIFACTORY_CREDENTIALS}") {
                 artifactoryImageTag.push()
               }
             }
@@ -160,11 +216,15 @@ pipeline {
     always {
       archiveArtifacts artifacts: '*.txt', onlyIfSuccessful: true
       sh "docker rmi ${DOCKERHUB_IMAGE_NAME}:latest"
-      sh "docker rmi ${DOCKERHUB_IMAGE_NAME}:${DOCKER_IMAGE_TAG}"
-      sh "docker rmi registry.hub.docker.com/${DOCKERHUB_IMAGE_NAME}:latest"
-      sh "docker rmi registry.hub.docker.com/${DOCKERHUB_IMAGE_NAME}:${DOCKER_IMAGE_TAG}"
-      sh "docker rmi ${ARTIFACTORY_IMAGE_NAME}:latest"
-      sh "docker rmi ${ARTIFACTORY_IMAGE_NAME}:${DOCKER_IMAGE_TAG}"
+      //sh "docker rmi ${DOCKERHUB_IMAGE_NAME}:${DOCKER_IMAGE_TAG}"
+    }
+    success {
+      script {
+        if (params.PUSH_ARTIFACTORY == true) {
+          sh "docker rmi ${ARTIFACTORY_IMAGE_NAME}:latest"
+          sh "docker rmi ${ARTIFACTORY_IMAGE_NAME}:${DOCKER_IMAGE_TAG}"
+        }
+      }
     }
   }
 }
